refactor: APIファイルをリソース別に再構成
- sync.goからメソッドを各リソースファイルに移動
- GetProjects, DeleteProjectSync → projects.go
- GetItems, CompleteItem, DeleteItem → tasks.go
- GetSections, GetAllSections → 新しいsections.go
- 空になったsync.goを削除

これによりAPIの構成がリソース別に整理され、
どこに何があるかが直感的にわかりやすくなった。
